requirements.txt
Flask
numpy
pandas
python-dotenv
Requests
torch==1.13.0
tqdm
gunicorn
